+++ CHANGES.md
@@ -13,6 +13,7 @@
   matched everywhere) - new `isRealSafari` (`src/utils/media.js`)
   properly excludes Chromium browsers
 - fix(search): missing error types selection for special Errors
+- fix(search): avoid inconsistent selectors across types
 - fix(demo): add Boolean, Number, String objects
 - test: improve coverage